requirements.txt
aiogram>=3.7.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
